requirements.txt
aiohttp==3.9.1
websockets==12.0
pytz==2024.1
python-dotenv==1.0.0
pyyaml==6.0.1
typing-extensions==4.9.0
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
aioresponses==0.7.6
pytest-mock==3.12.0 